[Improve] Enhanced stability of Kudu E2E
