+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-kudu-e2e/src/test/resources/fake_to_kudu_with_multipletable.conf
@@ -82,6 +82,6 @@ source {
 
 sink {
    kudu{
-    kudu_masters = "kudu-master-multiple:7051"
+    kudu_masters = "kudu-master:7051"
  }
 }
