+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-kudu-e2e/src/test/resources/kudu_to_assert_with_multipletable.conf
@@ -27,7 +27,7 @@ env {
 source {
   # This is a example source plugin **only for test and demonstrate the feature source plugin**
   kudu{
-   kudu_masters = "kudu-master-multiple:7051"
+   kudu_masters = "kudu-master:7051"
    table_list = [
    {
     table_name = "kudu_source_table_1"
