+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-kudu-e2e/src/test/resources/write-cdc-changelog-to-kudu.conf
@@ -68,7 +68,7 @@ source {
 
 sink {
    kudu{
-    kudu_masters = "kudu-master-cdc:7051"
-    table_name = "kudu_sink_table"
+    kudu_masters = "kudu-master:7051"
+    table_name = "kudu_cdc_sink_table"
  }
 }
